ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
opencv-python==4.11.0.86
numpy>=2.2.0
tensorflow==2.21.0
pillow==11.1.0
